+++ modules/tinymce/src/core/main/ts/DragDropOverrides.ts
@@ -1,3 +1,4 @@
+import { DataTransfer, DataTransferContent } from '@ephox/dragster';
 import { Arr, Optional, Singleton, Throttler, Type } from '@ephox/katamari';
 import { SugarElement } from '@ephox/sugar';
 
@@ -33,6 +34,7 @@ const mouseRangeToTriggerScrollOutsideEditor = 16;
 
 interface State {
   element: HTMLElement;
+  dataTransfer: DataTransfer;
   dragging: boolean;
   screenX: number;
   screenY: number;
@@ -60,12 +62,6 @@ const isValidDropTarget = (editor: Editor, targetElement: Node | null, dragEleme
   } else {
     return editor.dom.isEditable(targetElement);
   }
-};
-
-const cloneElement = (elm: HTMLElement) => {
-  const cloneElm = elm.cloneNode(true) as HTMLElement;
-  cloneElm.removeAttribute('data-mce-selected');
-  return cloneElm;
 };
 
 const createGhost = (editor: Editor, elm: HTMLElement, width: number, height: number) => {
@@ -243,6 +239,7 @@ const start = (state: Singleton.Value<State>, editor: Editor) => (e: EditorEvent
 
       state.set({
         element: ceElm,
+        dataTransfer: DataTransfer.createDataTransfer(),
         dragging: false,
         screenX: e.screenX,
         screenY: e.screenY,
@@ -274,6 +271,17 @@ const placeCaretAt = (editor: Editor, clientX: number, clientY: number) => {
   );
 };
 
+const dispatchDragEvent = (editor: Editor, type: DragEvents.DragEventType, target: Element, dataTransfer: DataTransfer, mouseEvent?: EditorEvent<MouseEvent>): EditorEvent<DragEvent> => {
+  if (type === 'dragstart') {
+    DataTransferContent.setHtmlData(dataTransfer, editor.dom.getOuterHTML(target));
+  }
+
+  const event = DragEvents.makeDragEvent(type, target, dataTransfer, mouseEvent);
+  const args = editor.dispatch(type, event);
+
+  return args;
+};
+
 const move = (state: Singleton.Value<State>, editor: Editor) => {
   // Reduces laggy drag behavior on Gecko
   const throttledPlaceCaretAt = Throttler.first((clientX: number, clientY: number) => placeCaretAt(editor, clientX, clientY), 0);
@@ -284,7 +292,12 @@ const move = (state: Singleton.Value<State>, editor: Editor) => {
     const movement = Math.max(Math.abs(e.screenX - state.screenX), Math.abs(e.screenY - state.screenY));
 
     if (!state.dragging && movement > 10) {
-      const args = editor.dispatch('dragstart', DragEvents.makeDragstartEventFromMouseEvent(e, state.element));
+      const args = dispatchDragEvent(editor, 'dragstart', state.element, state.dataTransfer, e);
+      // TINY-9601: dataTransfer is writable in dragstart, so keep it up-to-date
+      if (Type.isNonNullable(args.dataTransfer)) {
+        state.dataTransfer = args.dataTransfer;
+      }
+
       if (args.isDefaultPrevented()) {
         return;
       }
@@ -320,21 +333,22 @@ const drop = (state: Singleton.Value<State>, editor: Editor) => (e: EditorEvent<
     state.intervalId.clear();
     if (state.dragging) {
       if (isValidDropTarget(editor, getRawTarget(editor.selection), state.element)) {
-        const targetClone = cloneElement(state.element);
         const dropTarget = editor.getDoc().elementFromPoint(e.clientX, e.clientY) ?? editor.getBody();
-        const args = editor.dispatch('drop', DragEvents.makeDropEventFromMouseEvent(e, dropTarget));
+        const args = dispatchDragEvent(editor, 'drop', dropTarget, state.dataTransfer, e);
 
         if (!args.isDefaultPrevented()) {
           editor.undoManager.transact(() => {
             removeElementWithPadding(editor.dom, state.element);
-            editor.insertContent(editor.dom.getOuterHTML(targetClone));
+            // TINY-9601: Use dataTransfer property to determine inserted content on drop. This allows users to
+            // manipulate drop content by modifying dataTransfer in the dragstart event.
+            DataTransferContent.getHtmlData(state.dataTransfer).each((content) => editor.insertContent(content));
             editor._selectionOverrides.hideFakeCaret();
           });
         }
       }
 
       // Use body as the target since the element we are dragging no longer exists. Native drag/drop works in a similar way.
-      editor.dispatch('dragend', DragEvents.makeDragendEventFromMouseEvent(e, editor.getBody()));
+      dispatchDragEvent(editor, 'dragend', editor.getBody(), state.dataTransfer, e);
     }
   });
 
@@ -345,11 +359,10 @@ const stopDragging = (state: Singleton.Value<State>, editor: Editor, e: Optional
   state.on((state) => {
     state.intervalId.clear();
     if (state.dragging) {
-      const event = e.fold(
-        () => DragEvents.makeDragendEvent(state.element),
-        (mouseEvent) => DragEvents.makeDragendEventFromMouseEvent(mouseEvent, state.element)
+      e.fold(
+        () => dispatchDragEvent(editor, 'dragend', state.element, state.dataTransfer),
+        (mouseEvent) => dispatchDragEvent(editor, 'dragend', state.element, state.dataTransfer, mouseEvent)
       );
-      editor.dispatch('dragend', event);
     }
   });
   removeDragState(state);