+++ modules/dragster/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- New `isActive` api to Dragger to check if it's active or not.
+
 ## 7.0.0 - 2022-03-03
 
 ### Changed